Rename folder 'signupTrip' to 'trips/signup'
